Allow LOF, LOT to start on even or odd page
Previous beavtex.cls forced the list of figures to begin on a right side page. If adding a flypage or multipage abstract, this can cause unwanted blank pages between list of figures and list of tables, etc.
Also added blank page before abstract to conform with the graduate school guidelines as of 5/2015.